Corrections

+++ Test/src/test/java/be/groups/glanguage/glanguage/api/entities/formula/implementations/binary/FormulaModuloTest.java
@@ -98,7 +98,7 @@ public class FormulaModuloTest extends BaseDatabaseTest {
 		FormulaModulo formula =
 				new FormulaModulo(FormulaDescriptionFactory.getDescription(FormulaType.OP_MODULO), numerator, denominator);
 				
-		assertEquals(FormulaReturnType.NUMERIC, formula.getReturnType());
+		assertEquals(FormulaReturnType.INTEGER, formula.getReturnType());
 	}
 	
 	/**
